# doc/api.yaml
openapi: 3.0.3
info:
  title: WASAphoto API
  description: |-
    This OpenAPI document describes WASAphoto functions
  version: "1"
...
security:
  - bearerAuth: []
paths:
...
  /profiles/{user_id}/bans/{target_uid}:
    parameters:
      - name: user_id
        in: path
        description: user's id that wants to ban/unban another user
        schema: {$ref: "#/components/schemas/UserID"}
        required: true
      - name: target_uid
        in: path
        description: user's id that is going to be banned/unbanned
        schema: {$ref: "#/components/schemas/UserID"}
        required: true
    put:
      tags: ["action"]
      operationId: banUser
      summary: ban a user
      description: add the target_uid user to user's current_uid banned list
      requestBody:
        content:
          application/json:
            schema: {$ref: "#/components/schemas/User" }
      responses:
        "200":
          description: user successfully added to ban list
          content:
            application/json:
              schema: {$ref: "#/components/schemas/User"}
        "400":
          description: Invalid request (e.g., invalid user_id or target_uid)
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
              example:
                error: "Invalid user_id or target_uid"
        "404":
          description: User not found
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
              example:
                error: "User not found"
        "500":
          description: Server error
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
              example:
                error: "Internal server error"
    delete:
      tags: ["action"]
      operationId: unbanUser
      summary: unban a user
      description: remove the target_uid user from user's current_uid banned list
      responses:
        "204":
          description: user successfully unbanned
        "400":
          description: Invalid request (e.g., invalid user_id or target_uid)
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
              example:
                error: "Invalid user_id or target_uid"
        "404":
          description: User not found
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
              example:
                error: "User not found"
        "500":
          description: Server error
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
              example:
                error: "Internal server error"
    get:
      tags: ["action"]
      operationId: getBan
      summary: get the user who banned another user
      responses:
        "200": 
          description: user successfully retrieved
          content:
            application/json:
              schema: {$ref: "#/components/schemas/User"}
        "400":
          description: Invalid request (e.g., invalid user_id or target_uid)
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
              example:
                error: "Invalid user_id or target_uid"
        "404":
          description: User not found
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
              example:
                error: "User not found"
        "500":
          description: Server error
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
              example:
                error: "Internal server error"
...
  /profiles/{user_id}/photos:
    post:
      tags: ["action"]
      operationId: uploadPhoto
      summary: Upload a photo
      description: User uploads a photo to their profile
      parameters:
        - name: user_id
          in: path
          description: ID of the user who wants to upload a photo
          required: true
          schema:
            $ref: "#/components/schemas/UserID"
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema: { $ref:"#/components/schemas/UserPhoto"}
      responses:
        '201':
          description: Photo uploaded successfully
          content:
            application/json:
              schema: { $ref:"#/components/schemas/UserPhoto" }
        '400':
          description: Bad request
        '500':
          description: Internal server error
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    CommentID:
      title: Comment ID 
      description: id of a comment
      type: object
      properties:
        id:
          description: id of a comment
          type: integer
          example: 58
    CommentText:
      title: Comment Text
      description: a string representing a text under a post
      type: object
      properties:
        text:
          type: string
          description: a string representing a text under a post
          example: hey,nice photo
          pattern: '^[^\/\\]{1,64}$'
          maxLength: 100
          minLength: 1
    Comment:
      title: Comment
      description: "text added from users to posts"
      type: object
      properties:
        username:
          $ref: "#/components/schemas/Username"
        comment_id:
          $ref: "#/components/schemas/CommentID"
        comment_text:
          $ref: "#/components/schemas/CommentText"
        timestamp:
          type: string
          description: date and time when the comment is posted
          format: date-time
          example: "2010-01-20T16:10:19Z"
          minLength: 20
          maxLength: 20
          readOnly: true
    Liked:
      title: Liked list
      description: list of all liked photos
      type: object
      properties:
        photo_id:
          type: array
          description: list of all liked photos
          items: {$ref: "#/components/schemas/PhotoID"}
          minItems: 0
          maxItems: 10000000
...
    PhotoID:
      title: photo id
      description: "photo identificator"
      type: object
      properties:
        id:
          type: integer
          example: 45
    Followed:
      title: Followed list
      description: list of users' names and IDs followed by user
      type: object
      properties:
        user:
          type: array
          description: list of users' names and IDs followed by user
          items: 
            oneOf:
              - {$ref: "#/components/schemas/Username"}
              - {$ref: "#/components/schemas/UserID"}
          minItems: 0
          maxItems: 10000000
    Banned:
      title: Ban list
      description: list of users' names and IDs banned by user
      type: object
      properties:
        user:
          type: array
          description: list of users' names and IDs banned by user
          items:
            oneOf:
              - {$ref: "#/components/schemas/Username"}
              - {$ref: "#/components/schemas/UserID"}
          minItems: 0
          maxItems: 10000000
    UserPhoto:
      title: User Photo
      description: Details of an uploaded photo
      type: object
      properties:
        photo_id: {$ref: "#/components/schemas/PhotoID"}
        user_id: {$ref: "#/components/schemas/UserID"}
        image:
          $ref: '#/components/schemas/Image'
        timestamp:
          type: string
          description: date and time when the image was posted
          format: date-time
          example: "2010-09-20T16:10:19Z"
          minLength: 20
          maxLength: 20
          readOnly: true
        likeCount:
          description: number of likes that the photo received
          type: integer
          example: 10
          minimum: 0
          readOnly: true
        commentCount:
          description: number of comments that the photo received
          type: integer
          example: 10
          minimum: 0
          readOnly: true
        liked:
          description: flag that retrieves if the users liked to the photo
          type: boolean
          example: false
          readOnly: true
    # UserPhoto:
    #   title: User photo
    #   description: photo uploaded by the user
    #   properties:
    #     user_id: {$ref: "#/components/schemas/UserID"}
    #     pid: {$ref: "#/components/schemas/PhotoID"}
    #     photo: {$ref: "#/components/schemas/Image"}
#          title: photo path
#          description: "describes the path of a photo"
#          type : string
#          pattern: '^\/(?:[\w]+\/)*[\w]+\.[a-zA-Z]{2,}$'
#          example: "path/images/photo.png"
#          minLength: 1
#          maxLength: 100
          
        caption: {$ref: "#/components/schemas/Caption"}
          
#        likes_number:
#          title: likes number
#          description: "number of likes received from users"
#          type: integer
#          example: 500
#        comments_number:
#          title: comments number
#          description: "number of comments received from users"
#          type: integer
#          example: 5
#        comments_list:
#          title: comments list
#          description: array of Comment
#          type: array
#          items: {$ref: "#/components/schemas/Comment"}
#          minItems: 0
#          maxItems: 10000000
        time:
          title: time
          description: pubblication time
          type: string
          format: date-time
          pattern: "^[0-9]{4}-[0-9]{2}-[0-9]{2}T[0-9]{2}:[0-9]{2}:[0-9]{2}$"
          example: "2023-11-21T23:19:59"
          minLength: 19
          maxLength: 19
    Image: 
          title: photo
          description: base 64 image.
          type: string
          format: byte
          minLength: 0
          maxLength: 1000000000
    Caption:
          description: "description of a photo"
          type: string
          pattern: '^[^\/\\]{1,64}$'
          example: "look at me!"
          minLength: 0
          maxLength: 50
...
    UserID:
      title: User ID
      description: "this is the user id"
      type: object
      properties:
        id:
          description: "user identificator number"
          type: integer
          example: 1
#          readOnly: true
    Username:
      title: Username
      description: "this is the user's name"
      type: object
      properties:
        username:
          type: string
          description: "name of the user"
          example: "Simone"
          pattern: '^[a-z][a-z0-9]*$'
          minLength: 3
          maxLength: 16
    User:
      title: User
      description: "all informations about User"
      type: object
      properties:
        uid:
          $ref: "#/components/schemas/UserID"
        username:
          $ref: "#/components/schemas/Username"
#        follow_list:
#          $ref: "#/components/schemas/Followed"
#        ban_list:
#          $ref: "#/components/schemas/Banned"
#        liked_list:
#          $ref: "#/components/schemas/Liked"
...